refactor: Result 컴포넌트로 변경
